+++ lib/themes/dark_theme.dart
@@ -1,29 +1,24 @@
 import 'package:flutter/material.dart';
-
+import 'package:flutter/services.dart';
 import 'package:game_list/themes/nord.dart';
 
 ThemeData darkTheme = ThemeData(
-  fontFamily: 'Inter',
+  brightness: Brightness.dark,
+  primaryColor: Nord.nord1,
   accentColor: Nord.nord11,
+  fontFamily: 'Inter',
+  scaffoldBackgroundColor: Nord.nord1,
+  cardColor: Nord.nord3,
+  toggleableActiveColor: Nord.nord11,
   appBarTheme: AppBarTheme(
-    color: Nord.nord0,
-    brightness: Brightness.dark,
+    shadowColor: Colors.transparent,
+    systemOverlayStyle: SystemUiOverlayStyle(
+        statusBarColor: Colors.transparent,
+        systemNavigationBarColor: Nord.nord0,
+        systemNavigationBarIconBrightness: Brightness.light),
   ),
   bottomNavigationBarTheme: BottomNavigationBarThemeData(
       backgroundColor: Nord.nord0,
       selectedItemColor: Nord.nord11,
       unselectedItemColor: Nord.nord6),
-  brightness: Brightness.dark,
-  cardColor: Nord.nord2,
-  primaryTextTheme: TextTheme(
-      headline6: TextStyle(
-    color: Nord.nord6,
-  )),
-  scaffoldBackgroundColor: Nord.nord1,
-  textTheme: TextTheme(
-    bodyText2: TextStyle(
-      color: Nord.nord6,
-    ),
-  ),
-  toggleableActiveColor: Nord.nord11,
 );