+++ lib/themes/light_theme.dart
@@ -1,29 +1,24 @@
 import 'package:flutter/material.dart';
-
+import 'package:flutter/services.dart';
 import 'package:game_list/themes/nord.dart';
 
 ThemeData lightTheme = ThemeData(
-  fontFamily: 'Inter',
+  brightness: Brightness.light,
+  primaryColor: Nord.nord5,
   accentColor: Nord.nord11,
+  fontFamily: 'Inter',
+  scaffoldBackgroundColor: Nord.nord5,
+  cardColor: Nord.nord6,
+  toggleableActiveColor: Nord.nord11,
   appBarTheme: AppBarTheme(
-    color: Nord.nord0,
-    brightness: Brightness.dark,
+    shadowColor: Colors.transparent,
+    systemOverlayStyle: SystemUiOverlayStyle(
+        statusBarColor: Colors.transparent,
+        systemNavigationBarColor: Nord.nord4,
+        systemNavigationBarIconBrightness: Brightness.dark),
   ),
   bottomNavigationBarTheme: BottomNavigationBarThemeData(
       backgroundColor: Nord.nord4,
       selectedItemColor: Nord.nord11,
       unselectedItemColor: Nord.nord0),
-  brightness: Brightness.light,
-  cardColor: Nord.nord6,
-  primaryTextTheme: TextTheme(
-      headline6: TextStyle(
-    color: Nord.nord6,
-  )),
-  scaffoldBackgroundColor: Nord.nord5,
-  textTheme: TextTheme(
-    bodyText2: TextStyle(
-      color: Nord.nord0,
-    ),
-  ),
-  toggleableActiveColor: Nord.nord11,
 );